cleans rights management
