--- src/main/core/Entity/RoleOptions.php
@@ -1,71 +1,0 @@
-<?php
-
-/*
- * This file is part of the Claroline Connect package.
- *
- * (c) Claroline Consortium <consortium@claroline.net>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Claroline\CoreBundle\Entity;
-
-use Doctrine\ORM\Mapping as ORM;
-
-/**
- * @ORM\Table(name="claro_role_options")
- * @ORM\Entity()
- */
-class RoleOptions
-{
-    /**
-     * @ORM\Id
-     * @ORM\Column(type="integer")
-     * @ORM\GeneratedValue(strategy="AUTO")
-     */
-    protected $id;
-
-    /**
-     * @ORM\OneToOne(
-     *     targetEntity="Claroline\CoreBundle\Entity\Role"
-     * )
-     * @ORM\JoinColumn(name="role_id", onDelete="CASCADE")
-     */
-    protected $role;
-
-    /**
-     * @ORM\Column(type="json_array", nullable=true)
-     */
-    protected $details;
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function setId($id)
-    {
-        $this->id = $id;
-    }
-
-    public function getRole()
-    {
-        return $this->role;
-    }
-
-    public function setRole(Role $role)
-    {
-        $this->role = $role;
-    }
-
-    public function getDetails()
-    {
-        return $this->details;
-    }
-
-    public function setDetails($details)
-    {
-        $this->details = $details;
-    }
-}
