+++ src/plugin/exo/Tests/Controller/Api/ExerciseControllerTest.php
@@ -58,7 +58,7 @@ class ExerciseControllerTest extends TransactionalTestCase
         $roleManager = $this->client->getContainer()->get('claroline.manager.role_manager');
 
         // add open permissions to all users
-        $rightsManager->editPerms(1, $roleManager->getRoleByName('ROLE_USER'), $this->exercise->getResourceNode());
+        $rightsManager->update(1, $roleManager->getRoleByName('ROLE_USER'), $this->exercise->getResourceNode());
 
         $this->om->flush();
     }
